requirements.txt.txt
python-telegram-bot==22.5
requests
beautifulsoup4
apscheduler
pytz
